Migrate to standalone components: step 2-3: Bootstrap the project using standalone APIs. app works fine but several failing tests
